WORKING 64 BIT(QEMU)
YAYYYYYYYYYYYYYYYYYYYYYYYYYYYY

+++ graphics/fb.h
@@ -7,7 +7,7 @@ typedef struct {
     int available;
 } framebuffer_t;
 extern framebuffer_t fb;
-void fb_init(uint32_t magic, uint32_t mbi);
+void fb_init(uint64_t magic, uint64_t mbi);
 void fb_put_pixel(uint32_t x, uint32_t y, uint32_t color);
 void fb_fill_rect(uint32_t x, uint32_t y, uint32_t w, uint32_t h, uint32_t color);
 void fb_clear(uint32_t color);
